% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as}
\alias{as}
\title{Change into a chosen type}
\usage{
as(x, type)
}
\arguments{
\item{x}{object to be coerced}

\item{type}{type to be coerced into

The "type" parameter takes one of the three values: "integer", "double" 
or "character".}
}
\value{
the same objected with changed type
}
\description{
Change into a chosen type
}
\examples{
as(1:3, "character")
}
