% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ircurve.R
\name{as_ircurve}
\alias{as_ircurve}
\title{Create an interest rates curve}
\usage{
as_ircurve(rates = double(), type = "forward", period = "year")
}
\arguments{
\item{rates}{vector of rates}

\item{type}{"forward" or "spot"}

\item{period}{"year" or "month"}
}
\value{
Interest rates curve, an object of class ircurve.
}
\description{
Creates an object of class ircurve which represents in interest rates curve.
The curve cosists of interest rates and is of type either "forward" or
"spot". The interest rates curve can be for monthly or yearly periods.
}
\examples{
as_ircurve(c(0.01, 0.015))
as_ircurve(c(0.002, 0.001), type = "spot", period = "month")
}
