% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ircurve.R
\name{change_period}
\alias{change_period}
\title{Change the period of an interest rates curve}
\usage{
change_period(ircurve, to)
}
\arguments{
\item{ircurve}{an object of class ircurve}

\item{to}{"year" or "month"}
}
\value{
An object of class ircurve with changed type.
}
\description{
Changes the period of an interest rates curve (object of class ircurve)
from year to monthl or vice versa.
}
\examples{
my_ircurve <- as_ircurve(rep(0.01, 12), period = "month")
change_period(my_ircurve, to = "year")
}
