% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ircurve.R
\name{change_type}
\alias{change_type}
\title{Change the type of an interest rates curve}
\usage{
change_type(ircurve, to)
}
\arguments{
\item{ircurve}{an object of class ircurve}

\item{to}{"forward" or "spot"}
}
\value{
An object of class ircurve with changed type.
}
\description{
Changes the type of an interest rates curve (object of class ircurve) 
from forward to spot or vice versa.
}
\examples{
my_ircurve <- as_ircurve(c(0.1, 0.2, 0.3), type = "spot")
change_type(my_ircurve, to = "forward")
}
