% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_to_character}
\alias{coerce_to_character}
\title{Coerce factors/POSIXct/Dates into characters}
\usage{
coerce_to_character(tables)
}
\arguments{
\item{tables}{a list with two elements: x and y}
}
\value{
a list with two elements: x and y
}
\description{
Coerce factors/POSIXct/Dates into characters
}
\keyword{internal}
