% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two tables}
\usage{
compare(x, y)
}
\arguments{
\item{x}{the first data frame}

\item{y}{the second data frame}
}
\value{
data frame
}
\description{
Returns the effect of comparison of the two tables. It gets common columns
and number of rows of the two tables. In case of type mismatches, 
it coerces the weaker type into a stronger type. The output contains 
the absolute difference for numerical values and the_same/different for
characters.
}
\examples{
x <- data.frame(
  a = rep(1, 3),
  b = rep(2, 3)
)
y <- data.frame(
  a = rep(2, 3),
  b = rep(2, 3)
)
compare(x, y)
}
