% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_utils.R
\name{.as}
\alias{.as}
\title{Change type}
\usage{
.as(x, type)
}
\arguments{
\item{x}{object to be coerced}

\item{type}{type to be coerced into ("integer", "double" or "character")}
}
\value{
the same objected with changed type
}
\description{
Change type
}
\examples{
.as(1:3. "character")
}
