% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_utils.R
\name{.getCommonColumns}
\alias{.getCommonColumns}
\title{Get common columns across two tables}
\usage{
.getCommonColumns(tables)
}
\arguments{
\item{tables}{a list with two elements: x and y}
}
\value{
a list with two elements: x and y
}
\description{
Get common columns across two tables
}
