% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_utils.R
\name{.getCommonNrRows}
\alias{.getCommonNrRows}
\title{Get common number of rows across two tables}
\usage{
.getCommonNrRows(tables)
}
\arguments{
\item{tables}{a list with two elements: x and y}
}
\value{
a list with two elements: x and y
}
\description{
Get common number of rows across two tables
}
