% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_utils.R
\name{.getCommonTypes}
\alias{.getCommonTypes}
\title{Get common types of columns across two tables}
\usage{
.getCommonTypes(tables)
}
\arguments{
\item{tables}{a list with two elements: x and y}
}
\value{
a list with two elements: x and y
}
\description{
Get common types of columns across two tables
}
