% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref.R
\name{dref_fdom}
\alias{dref_fdom}
\title{First day of a month}
\usage{
dref_fdom(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the first day of a month in reference to the base date.
}
\examples{
dref_fdom("2020-09-21")
}
