% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_fdoq.R
\name{dref_fdoq}
\alias{dref_fdoq}
\title{First day of a quarter}
\usage{
dref_fdoq(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the first day of a quarter in reference to the base date.
}
\examples{
dref_fdoq("2020-09-21")
}
