% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref.R
\name{dref_fdoy}
\alias{dref_fdoy}
\title{First day of an year}
\usage{
dref_fdoy(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the first day of a year in reference to the base date.
}
\examples{
dref_fdoy("2020-09-21")
}
