% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_fwdom.R
\name{dref_fwdom}
\alias{dref_fwdom}
\title{First working day of a month}
\usage{
dref_fwdom(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the first working day of a month in reference to the base date.
}
\examples{
dref_fwdom("2020-11-21")
}
