% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_fwdoq.R
\name{dref_fwdoq}
\alias{dref_fwdoq}
\title{First working day of a quarter}
\usage{
dref_fwdoq(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the first working day of a quarter in reference to the base date.
}
\examples{
dref_fwdoq("2020-09-21")
}
