% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_fwdoy.R
\name{dref_fwdoy}
\alias{dref_fwdoy}
\title{First working day of an year}
\usage{
dref_fwdoy(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the first working day of a year in reference to the base date.
}
\examples{
dref_fwdoy("2022-09-21")
}
