% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_ldom.R
\name{dref_ldom}
\alias{dref_ldom}
\title{Last day of a month}
\usage{
dref_ldom(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the last day of a month in reference to the base date.
}
\examples{
dref_ldom("2020-09-21")
}
