% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_ldoq.R
\name{dref_ldoq}
\alias{dref_ldoq}
\title{Last day of a quarter}
\usage{
dref_ldoq(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the last day of a month in reference to the base date.
}
