% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_ldoy.R
\name{dref_ldoy}
\alias{dref_ldoy}
\title{Last day of an year}
\usage{
dref_ldoy(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the last day of a year in reference to the base date.
}
\examples{
dref_ldoy("2020-09-21")
}
