% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_lwdom.R
\name{dref_lwdom}
\alias{dref_lwdom}
\title{Last working day of a month}
\usage{
dref_lwdom(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the last working day of a month in reference to the base date.
}
\examples{
dref_lwdom("2020-10-21")
}
