% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref.R
\name{dref_lwdoq}
\alias{dref_lwdoq}
\title{Last working day of a quarter}
\usage{
dref_lwdoq(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the last working day of a quarter in reference to the base date.
}
\examples{
dref_lwdoq("2020-09-21")
}
