% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_lwdoy.R
\name{dref_lwdoy}
\alias{dref_lwdoy}
\title{Last working day of an year}
\usage{
dref_lwdoy(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the last working day of a year in reference to the base date.
}
\examples{
dref_lwdoy("2022-09-21")
}
