% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_mtd.R
\name{dref_mtd}
\alias{dref_mtd}
\title{Month to date}
\usage{
dref_mtd(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the last day of the previous month.
}
\examples{
dref_mtd("2020-09-21")
}
