% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_qtd.R
\name{dref_qtd}
\alias{dref_qtd}
\title{Quarter to date}
\usage{
dref_qtd(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the last day of the previous quarter.
}
\examples{
dref_qtd("2020-09-21")
}
