% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dref_ytd.R
\name{dref_ytd}
\alias{dref_ytd}
\title{Year to date}
\usage{
dref_ytd(date)
}
\arguments{
\item{date}{base date in format YYYY-MM-DD}
}
\value{
date
}
\description{
Returns the last day of the previous year.
}
\examples{
dref_ytd("2020-09-21")
}
