% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_common_columns}
\alias{get_common_columns}
\title{Get common columns across two tables}
\usage{
get_common_columns(tables)
}
\arguments{
\item{tables}{a list with two elements: x and y}
}
\value{
a list with two elements: x and y
}
\description{
Get common columns across two tables
}
\keyword{internal}
