% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ircurve.R
\name{print.ircurve}
\alias{print.ircurve}
\title{Print ircurve}
\usage{
\method{print}{ircurve}(x, ...)
}
\arguments{
\item{x}{an object of class ircurve}

\item{...}{further arguments passed to or from other methods}
}
\value{
Prints the interest rates curve.
}
\description{
Print ircurve
}
\examples{
my_ircurve <- as_ircurve(c(0.022, 0.018))
print(my_ircurve)
}
